services/order-router/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.0
prometheus-client>=0.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.2.0
